requirements.txt
aspy.yaml==1.1.1
authy==2.2.2
autopep8==1.4
cached-property==1.5.1
certifi==2018.8.24
cfgv==1.1.0
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
flake8==3.5.0
gunicorn==19.9.0
identify==1.1.6
idna==2.7
mccabe==0.6.1
mongoengine==0.15.3
nodeenv==1.3.2
oauthlib==2.1.0
phonenumbers==8.9.14
pipenv==2018.7.1
pre-commit==1.3.0
psycopg2==2.7.5
pur==5.0.0
pycodestyle==2.4.0
pyflakes==1.6.0
PyJWT==1.6.4
pymongo==3.7.2
python-dotenv==0.9.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==1.7.0
social-auth-storage-mongoengine==1.0.1
toml==0.9.6
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
aioredis==1.2.0
asgiref==2.3.2
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
backcall==0.1.0
bleach==2.1.4
channels==2.1.5
channels-redis==2.3.1
cmmnbuild-dep-manager==2.2.2
colorama==0.3.9
constantly==15.1.0
cycler==0.10.0
daphne==2.2.3
decorator==4.3.0
django-cors-headers==2.4.0
entrypoints==0.2.3
hiredis==0.2.0
html5lib==1.0.1
hyperlink==18.0.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
isort==4.3.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.15.1
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
pip==9.0.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.1.1
pyparsing==2.2.0
python-dateutil==2.7.3
pyzmq==17.1.2
qtconsole==4.4.1
redis==2.10.6
Send2Trash==1.5.0
setuptools==40.4.2
simplegeneric==0.8.1
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
wrapt==1.10.11
zope.interface==4.6.0
